cts/requirements.txt
amqp==2.2.2
Babel==2.5.1
backports-abc==0.5
billiard==3.5.0.3
bootstrap-admin==0.3.7.1
celery==4.1.0
certifi==2017.4.17
chardet==3.0.4
crypto==1.4.1
Django==1.11.4
django-celery-beat==1.0.1
django-cors-headers==2.1.0
django-crispy-forms==1.6.1
django-debug-toolbar==1.8
django-endless-pagination==2.0
django-filter==1.0.4
djangorestframework==3.6.3
flower==0.9.2
futures==3.1.1
gevent==1.2.2
greenlet==0.4.12
html5lib==0.999999999
httpie==0.9.9
idna==2.5
JPype1==0.6.2
kombu==4.1.0
Markdown==2.5.2
meld3==1.0.2
MySQL-python==1.2.3
mysqlclient==1.3.12
Naked==0.1.31
olefile==0.44
Pillow==2.7.0
pyasn1==0.2.3
Pygments==2.2.0
python-memcached==1.58
python-openid==2.2.5
pytz==2017.2
PyYAML==3.12
requests==2.18.1
shellescape==3.4.1
singledispatch==3.4.0.3
six==1.11.0
South==1.0.2
sqlparse==0.2.4
supervisor==3.3.3
tornado==4.5.2
unicodecsv==0.14.1
urllib3==1.21.1
uWSGI==2.0.15
vine==1.1.4
webencodings==0.5.1
